pyproject.toml
[tool.poetry]
name = "nighthawk"
version = "0.1.0"
description = ""
authors = ["Jacek Chmielewski <jchmielewski@teonite.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
python-benedict = "^0.33.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
